- quests should work after merging.

+++ BattleOfTheMachines/BattleOfTheMachines.Data.Models/Quest.cs
@@ -20,7 +20,7 @@
         public string Description { get; set; }
 
         [Required]
-        public float DurationInMinutes { get; set; }
+        public int DurationInMinutes { get; set; }
 
         public byte[] Image { get; set; }
